move settings.json to defaults.json, add settings.js
