--- packages/app-spec-squad-chess/src/settings.json
@@ -1,12 +1,0 @@
-{
-  "metastoreWs": "wss://squad-mock-metastore.herokuapp.com/",
-  "matchmakingWs": "wss://squad-matchmaking.herokuapp.com",
-  "gameAddress": "8c875b7990077c166479df9ee2aafff7cc7bf82c54741ff47d169a56f18fe5d5",
-  "boardConfig": {
-    "squares": {
-      "size": 5,
-      "lightColor": "#e5b578",
-      "darkColor": "#b67028"
-    }
-  }
-}
